Rename module device_model -> device
