+++ archinstall/lib/disk/fido.py
@@ -4,7 +4,7 @@ import getpass
 from pathlib import Path
 from typing import ClassVar
 
-from archinstall.lib.models.device_model import Fido2Device
+from archinstall.lib.models.device import Fido2Device
 
 from ..exceptions import SysCallError
 from ..general import SysCommand, SysCommandWorker, clear_vt100_escape_codes_from_str
