+++ archinstall/lib/disk/subvolume_menu.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import assert_never, override
 
-from archinstall.lib.models.device_model import SubvolumeModification
+from archinstall.lib.models.device import SubvolumeModification
 from archinstall.lib.translationhandler import tr
 from archinstall.tui.curses_menu import EditMenu
 from archinstall.tui.result import ResultType
