+++ archinstall/lib/models/device.py
