+++ archinstall/scripts/guided.py
@@ -11,7 +11,7 @@ from archinstall.lib.global_menu import GlobalMenu
 from archinstall.lib.installer import Installer, accessibility_tools_in_use, run_custom_user_commands
 from archinstall.lib.interactions.general_conf import PostInstallationAction, ask_post_installation
 from archinstall.lib.models import Bootloader
-from archinstall.lib.models.device_model import (
+from archinstall.lib.models.device import (
 	DiskLayoutType,
 	EncryptionType,
 )
